% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame.}

\item{cl}{true class value of training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
Return a list of objects:
\itemize{
\item \code{class}: Vector of predicted class for all observations,
\item \code{cv_err}: Numeric with the cross-validation misclassification error.
}
}
\description{
This function predicts output from existed variables by using k-nearest
neighbors cross-validation
}
\examples{
data(my_penguins)
penguins_df <- na.omit(my_penguins)
my_knn_cv(penguins_df[ , 3:6], penguins_df$species, 1, 5)

}
\keyword{prediction}
