% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-Test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
List with elements,
\itemize{
\item \code{test_stat}: the numeric test statistic,
\item \code{df}: the degrees of freedom,
\item \code{alternative}: the value of the parameter,
\item \code{p_val}: the numeric p-value.
}
}
\description{
A function that performs one and two sample t-tests on vectors of data
}
\examples{
test = c(1:25)
my_t.test(test, alternative = "less", mu = 10)
my_t.test(test, alternative = "two.sided", mu = 10)

}
\keyword{inference}
